+++ packages/instant-apps-components/src/components/instant-apps-landing-page/support/enum.ts
@@ -1,0 +1,11 @@
+export enum AlignmentPositions {
+  TopLeft = 'top-left',
+  TopCenter = 'top-center',
+  TopRight = 'top-right',
+  CenterLeft = 'center-left',
+  Center = 'center',
+  CenterRight = 'center-right',
+  BottomLeft = 'bottom-left',
+  BottomCenter = 'bottom-center',
+  BottomRight = 'bottom-right',
+}