+++ packages/instant-apps-components/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.0.0-beta.125
+
+### instant-apps-landing-page
+
+- Update `alignment` values to a string value
+- Update `disableTransition` default to `true`
+
 ## v1.0.0-beta.124
 
 ### instant-apps-filter-list